+++ test/StkTableTest.vue
@@ -30,6 +30,7 @@
             v-model:columns="columns"
             :row-height="28"
             :header-row-height="36"
+            :hide-header-title="['age']"
             fixed-col-shadow
             row-key="name"
             :auto-resize="() => console.log('auto-resize')"